Group2_ScriptB_LinkedIn/requirements.txt
# requirements.txt

# Data Collection
requests>=2.31.0
playwright>=1.40.0
beautifulsoup4>=4.12.0

# NLP & Embeddings
openai>=1.0.0
sentence-transformers>=2.2.0
nltk>=3.8.0

# Clustering & Dimensionality Reduction
scikit-learn>=1.3.0
hdbscan>=0.8.33
umap-learn>=0.5.5

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0